Switch back to adding a script instead of a command to the pipeline.
More recent versions of Pash support adding a script but now do not
work if a command is passed. Previously passing the text
'Install-Package foo' as a command would have been executed by Pash.
Now this fails to resolve the command. Adding the text as a script
successfully executes the command.