Support -Version parameters in cmdlets when using Pash.
Pash does not currently support the command lines:

  Install-Package NUnit -Version 2.6.2
  Install-Package jquery -Version 1.7.1.2

The version number should be parsed by Pash as a string but currently
it parses it as 2 or 3 doubles. As a workaround the command line is
checked to see if it has the -Version parameter and if this exists
then the next parameter is surrounded by double quotes so that
Pash correctly converts this to a SemanticVersion.