Show assembly references in solution window.
Update the assembly references in the solution window after the
information is received from dnx. Currently only showing the
references for the first framework that the project supports.
SharpDevelop has no support for changing the current code
completion context in the text editor.

For an ASP.NET 5 web application with a lot of references the Solution
window flickers a lot when adding the references. Tried to add them in
one step but that makes no difference.