Restore NuGet packages if they are missing.
The dependencies in the Projects window will change as the NuGet
packages are restored when they are missing or if the project.json
file is modified.

Had some problems with missing NuGet packages when the project is
opened causing the PackagesRestoreTool to be repeatedly run. When the
first restore is triggered due to missing dependencies the dnu restore
modifies the project.lock.json which then triggers another dnu restore
to be started. So it ends up always running the dnu restore. To
workaround this the PackagesRestoreTool has been modified to not run
dnu restore if there is already a lock for that project which
hopefully should mean that a restore is currently being run for that
project.

Also added Dispose methods to the file system watcher and the project
system. Presumably in omnisharp another separate process is created
for the project system so it will be disposed along with the file
system watcher. Since the project system in this addin is not a
separate process the project system and file system watcher are
explicitly disposed.

Also handle null Data being null in the OutputDataReceived and the
ErrorDataReceived events of the Process class.

Also handle an IOException in the project system when trying to post
a message to the design time host when TriggerDependeees is called.
This exception was occuring before the file system watcher was being
disposed so it may be possible to revert this try catch.