Project templates can now have a fixed project directory.
The project template's solution descriptor can now defined a fixed
directory for all projects created.

<Options>
   <FixedProjectDirectory>src\${ProjectName}</FixedProjectDirectory>
</Options>

The directory is relative to the solution directory. The fixed
directory is used when creating a new solution or adding a project to
an existing solution.

In the New Project dialog if a project template is selected that has a
fixed project directory when creating a new solution then the create
solution directory check box is disabled. If such a project template
is selected when adding a new project to an existing solution then
the location text box and browse button are disabled.