Show DNX commands and runtimes in Debug menu.
All commands and runtimes in the project.json file for the startup
project are now shown in the Debug menu.

Debug
    Set DNX Command
        web
    Set DNX Runtime
        Default
        DNX 4.5.1
        DNX Core 5.0

The different commands and runtimes can be selected and the currently
selected ones will be shown with a check box.

Currently selecting a different runtimes does not alter which runtime
is used when running the application.