Run DNX app with selected runtime.
The selected runtime (CLR or CoreCLR) in the Debug menu is now
used when the application is run.

If the runtime is missing it will not currently be downloaded
automatically and needs to be already installed.

Debugging with CoreCLR does not work. The debugger times out and
then SharpDevelop is left in a bad state and throws exceptions.