Project references now reflect the selected DNX runtime target.
Changing the DNX runtime via the Debug menu will cause the Projects
window to update so the References node for the project shows the
references for the selected DNX framework.

If the default runtime is selected then the first DNX framework is
used as returned from the DNX host. This is probably incorrect.

Also using the DNX runtime to change the references is not the
best approach. Ideally a path selector in the text editor would be
used, in the same way it works in Visual Studio, so the DNX framework
selected there would determine the references being used.