Allow DNX project references to be removed.
Add a Remove context menu that is available when a project reference
dependency node is right clicked. Currently this will remove the
project reference completely from the project. Ideally it should only
remove it from the current DNX framework which is the parent of the
tree node. Another problem is that the dependency node completely
collapses when the dependencies are refreshed.