Fix library project with no references on DNX target framework changed
DNX library projects use different target frameworks: .NETPlatform and
.NETFramework. So changing a console project's DNX target framework to
DNXCore failed to find a match so the references were removed from the
project. Now these frameworks are mapped to their matching equivalent
framework. If there is no match then one of the project's frameworks
will be selected so code completion does not completely stop working.

Really the DNX target framework should not affect the current context
for the the text editor. There should be a way to select the active
framework context in the text editor in its toolbar.