Allow active framework to be selected for class libraries.
Changing the execution target no longer changes the current framework
for the DNX project. Previously the references would be changed and
the preprocessor symbols updated when the execution target is changed.

Now to change the current framework there is a right click menu in the
text editor called Active Framework. Inside this menu is a list of the
frameworks for the project. Selecting the framework will update the
project references and preprocessor symbols.