Fix package restore not running on single processor machines.
If the Environment.ProcessorCount was 1 then restoring the packages
for the DNX project would not happen. The solution window would show
the Restoring message which would never be cleared. The problem was
that if the processor count was 1 then the number of concurrent
restore requests that could be run was set to 0 and the restore
background task would be blocked by the semaphore.