Allow automatic DNX restore to be enabled or disabled.
Added a new option to preferences in the DNX - General section. There
is now check box that can be used to enable or disable automatic
restore of dependencies for DNX projects. This is a global setting,
similar to the existing NuGet automatic package restore setting, which
will enable or disable DNX restore for all projects.